Better evaluation, better ply order.
Better evaluation method. Evaluation only inside a rectangle that
covers all stones (with extra border of thickness one). Detects plies
that ought to be made (gives a win), or which must be made (to prevent
a loss). It used to be the case that, after trying the next-ply
computed in a previous iteration in the iterative deepener, the rest
next-plies are tried row wise; now they’re evaluated (one level down),
and then tried from best to worst. This gives a better ply order and
faster pruning.